+++ src/ml_tooling/plots.py
@@ -6,12 +6,12 @@ from typing import Tuple, Sequence, Union
 import matplotlib.pyplot as plt
 from sklearn.metrics import roc_auc_score, roc_curve, r2_score
 import numpy as np
+import pandas as pd
 import itertools
 from matplotlib.axes import Axes
-from sklearn.pipeline import Pipeline
 
 from . import metrics
-from .utils import DataType, _is_percent
+from .utils import DataType, _is_percent, get_scoring_func
 
 
 def plot_roc_auc(y_true: DataType, y_proba: DataType, title: str = None, ax: Axes = None) -> Axes:
@@ -201,6 +201,7 @@ def plot_feature_importance(importance: DataType,
                             labels: DataType,
                             values: bool = None,
                             title: str = None,
+                            x_label: str = None,
                             ax: Axes = None,
                             top_n: Union[int, float] = None,
                             bottom_n: Union[int, float] = None
@@ -216,6 +217,9 @@ def plot_feature_importance(importance: DataType,
 
     :param title:
         Plot title
+
+    :param x_label:
+        Plot x-axis label
 
     :param values:
         Add value labels to end of each bar
@@ -262,7 +266,8 @@ def plot_feature_importance(importance: DataType,
     ax.barh(labels, np.abs(importance))
     ax.set_title(title)
     ax.set_ylabel('Features')
-    ax.set_xlabel('Importance')
+    x_label = 'Importance' if x_label is None else x_label
+    ax.set_xlabel(x_label)
     if values:
         for i, (x, y) in enumerate(_generate_text_labels(ax, horizontal=True)):
             ax.text(x, y, f"{importance[i]:.2f}", va='center')
@@ -377,29 +382,44 @@ def _generate_text_labels(ax, horizontal=False, padding=0.005):
         yield x_value, y_value
 
 
-def _get_feature_importance(model) -> np.ndarray:
+def _get_feature_importance(viz, samples, seed=1337, n_jobs=1, verbose=0) -> pd.DataFrame:
     """
     Helper function for extracting importances.
-    Checks for coef_ or feature_importances_ on model
 
-    :param model:
-        A sklearn estimator
+    Parameters
+    ----------
+    viz : BaseVisualize
+        An instance of BaseVisualizer
 
-    :return:
-        array of importances
+    samples : None, int, float
+
+        None - Original data set i used. Not recommended for small data sets
+
+        float - A new smaller data set is made from resampling with
+                replacement form the original data set. Not recommended for small data sets.
+                Recommended for very large data sets.
+
+        Int - A new  data set is made from resampling with replacement form the original data.
+              samples sets the number of resamples. Recommended for small data sets
+               to ensure stable estimates of feature importance.
+
+    seed : int
+        Seed for random number generator for permutation.
+
+    Returns
+    -------
+
+    np.array
+        Decrease in score when permuting features
+
+    float
+        Baseline score without permutation
+
     """
-    if isinstance(model, Pipeline):
-        model = model.steps[-1][1]
+    model = viz._model
+    scorer = get_scoring_func(viz.default_metric)
+    train_x = viz._data.train_x.copy()
+    train_y = viz._data.train_y.copy()
 
-    if hasattr(model, 'feature_importances_'):
-        importance = model.feature_importances_
-
-    elif hasattr(model, 'coef_'):
-        importance = model.coef_
-        if importance.ndim > 1:
-            importance = importance[0]
-    else:
-        model_name = model.__class__.__name__
-        raise VizError(f"{model_name} does not have either coef_ or feature_importances_")
-
-    return importance
+    return metrics._permutation_importances(model, scorer, train_x, train_y, samples, seed, n_jobs,
+                                            verbose)